Add text to a tidyprompt

Add text to a prompt by adding a prompt_wrap() which will append the text to the before or after the current prompt text.

Usage

add_text(prompt, text, position = c("after", "before"), sep = "\n\n")

Arguments

prompt

A single string or a tidyprompt() object

text

Text to be added to the current prompt text

position

Where to add the text; either "after" or "before".

sep

Separator to be used between the current prompt text and the text to be added

Value

A tidyprompt() with an added prompt_wrap() which will append the text to the end of the current prompt text

See also

Examples

prompt <- "Hi there!" |>
  add_text("How is your day?")
prompt
#> $base_prompt
#> [1] "Hi there!"
#> 
#> $prompt_wraps
#> $prompt_wraps[[1]]
#> $type
#> [1] "unspecified"
#> 
#> $modify_fn
#> function (original_prompt_text) 
#> {
#>     if (position == "after") {
#>         paste(original_prompt_text, text, sep = sep)
#>     }
#>     else {
#>         paste(text, original_prompt_text, sep = sep)
#>     }
#> }
#> <bytecode: 0x5648129d5738>
#> <environment: 0x5648129d6228>
#> 
#> $extraction_fn
#> NULL
#> 
#> $validation_fn
#> NULL
#> 
#> attr(,"class")
#> [1] "prompt_wrap"
#> 
#> 
#> attr(,"class")
#> [1] "tidyprompt"
prompt |>
  construct_prompt_text()
#> [1] "Hi there!\n\nHow is your day?"