Set ReAct mode for a prompt

This function enables ReAct mode for the evaluation of a prompt or a tidyprompt(). In ReAct mode, the large language model (LLM) is asked to think step by step, each time detailing a thought, action, and observation, to eventually arrive at a final answer. It is hypothesized that this may increase LLM performance at solving complex tasks. ReAct mode is inspired by the method described in Yao et al. (2022).

Usage

answer_by_react(prompt, extract_from_finish_brackets = TRUE)

Arguments

prompt

A single string or a tidyprompt() object

extract_from_finish_brackets

A logical indicating whether the final answer should be extracted from the text inside the "FINISH..." brackets.

Value

A tidyprompt() with an added prompt_wrap() which will ensure that the LLM follows the ReAct mode in answering the prompt

Details

Please note that ReAct mode may be most useful if in combination with tools that the LLM can use. See, for example, 'add_tools()' for enabling R function calling, or, for example, 'answer_as_code()' with 'output_as_tool = TRUE' for enabling R code evaluation as a tool.

References

Yao, S., Wu, Y., Cheung, W., Wang, Z., Narasimhan, K., & Kong, L. (2022). ReAct: Synergizing Reasoning and Acting in Language Models. doi:10.48550/arXiv.2210.03629

See also