Create or validate chat_history object

This function creates and validates a chat_history object, ensuring that it matches the expected format with 'role' and 'content' columns. It has separate methods for data.frame and character inputs and includes a helper function to add a system prompt to the chat history.

Usage

chat_history(chat_history)

Arguments

chat_history

A single string, a data.frame with 'role' and 'content' columns, or NULL. If a data.frame is provided, it should contain 'role' and 'content' columns, where 'role' is either 'user', 'assistant', or 'system', and 'content' is a character string representing a chat message

Value

A valid chat history data.frame (of class chat_history)

Examples

chat <- "Hi there!" |>
  chat_history()
chat
#>   role   content tool_result
#> 1 user Hi there!       FALSE

chat_from_df <- data.frame(
  role = c("user", "assistant"),
  content = c("Hi there!", "Hello! How can I help you today?")
) |>
  chat_history()
chat_from_df
#>        role                          content tool_result
#> 1      user                        Hi there!       FALSE
#> 2 assistant Hello! How can I help you today?       FALSE

# `add_msg_to_chat_history()` may be used to add messages to a chat history
chat_from_df <- chat_from_df |>
  add_msg_to_chat_history("Calculate 2+2 for me, please!")
chat_from_df
#>        role                          content tool_result
#> 1      user                        Hi there!       FALSE
#> 2 assistant Hello! How can I help you today?       FALSE
#> 3      user    Calculate 2+2 for me, please!       FALSE

# You can also call `add_msg_to_chat_history()` on a list
#   containing '$chat_history`;
#   e.g., a result of `send-prompt()` with `return_mode = "full"`
#     This may be used to continue a chat history in a new session
if (FALSE) { # \dontrun{
  result <- "Hi there!" |>
    send_prompt(return_mode = "full")
  result$response
} # }