Generic function to assist in creating llm_provider objects

This function can be used to create new llm_provider objects with different implementations of the complete_chat function.

Usage

create_llm_provider(complete_chat_function, parameters = list())

Arguments

complete_chat_function

Function that will be called by the llm_provider to complete a chat.

This function should take a chat_history dataframe as input (see ?validate_chat_history), and return a list of 'role' and 'content' for the next message (e.g., list(role = "user", content = "Hello")).

An llm_provider object will wrap the provided complete_chat_function with a validation of chat_history, also turning a single string into a valid chat_history dataframe. The provided function thus does not need to do this but should assume that the input is a valid chat_history dataframe.

Parameters passed in the parameters argument may be accessed by the complete_chat_function; this may be used to, for instance, store an API key, the name of the model to use, or other settings.

parameters

A named list of parameters that will be attached to the llm_provider object. These parameters can be used to configure the llm_provider. For instance, they can be used to store a model's name, an API key, or an endpoint. E.g., list(model = "my-llm-model", api_key = "my-api-key").

Value

A new llm_provider object