Get prompt wraps from a tidyprompt object

Get prompt wraps from a tidyprompt object

Usage

get_prompt_wraps(x, order = c("default", "modification", "evaluation"))

Arguments

x

A tidyprompt object

order

The order to return the wraps. Options are:

"default": as originally added to the object

"modification": as ordered for modification of the base prompt; ordered by type: check, unspecified, mode, tool, break. This is the order in which prompt wraps are applied during construct_prompt_text()

"evaluation": ordered for evaluation of the LLM response; ordered by type: tool, mode, break, unspecified, check. This is the order in which wraps are applied to the LLM output during send_prompt()

Value

A list of prompt wrap objects (see prompt_wrap())

See also

Examples

tidyprompt("Hi!")
#> <tidyprompt>
#> The base prompt is not modified by prompt wraps:
#> > Hi! 
#> Use 'x$base_prompt' to show the base prompt text.
#> Use 'x$construct_prompt_text()' to get the full prompt text.
#> 

# Add to a tidyprompt using a prompt wrap:
tidyprompt("Hi!") |>
  add_text("How are you?")
#> <tidyprompt>
#> The base prompt is modified by a prompt wrap, resulting in:
#> > Hi!
#> > 
#> > How are you? 
#> Use 'x$base_prompt' to show the base prompt text.
#> Use 'x$construct_prompt_text()' to get the full prompt text.
#> Use 'get_prompt_wraps(x)' to show the prompt wraps.
#> 

# Strings can be input for prompt wraps; therefore,
#   a call to tidyprompt() is not necessary:
"Hi" |>
  add_text("How are you?")
#> <tidyprompt>
#> The base prompt is modified by a prompt wrap, resulting in:
#> > Hi
#> > 
#> > How are you? 
#> Use 'x$base_prompt' to show the base prompt text.
#> Use 'x$construct_prompt_text()' to get the full prompt text.
#> Use 'get_prompt_wraps(x)' to show the prompt wraps.
#> 

# Example of adding extraction & validation with a prompt_wrap():
prompt <- "Hi" |>
  add_text("What is 5 + 5?") |>
  answer_as_integer()

if (FALSE) { # \dontrun{
  # tidyprompt objects are evaluated by send_prompt(), which will
  #   handle construct the prompt text, send it to the LLM provider,
  #   and apply the extraction and validation functions from the tidyprompt object
  prompt |>
    send_prompt(llm_provider_ollama())
  # --- Sending request to LLM provider (llama3.1:8b): ---
  #   Hi
  #
  #   What is 5 + 5?
  #
  #   You must answer with only an integer (use no other characters).
  # --- Receiving response from LLM provider: ---
  #   10
  # [1] 10

  # See prompt_wrap() and send_prompt() for more details
} # }

# `tidyprompt` objects may be validated with these helpers:
is_tidyprompt(prompt) # Returns TRUE if input is a valid tidyprompt object
#> [1] TRUE

# Get base prompt text
prompt$base_prompt
#> [1] "Hi"

# Get all prompt wraps
prompt$get_prompt_wraps()
#> [[1]]
#> $type
#> [1] "unspecified"
#> 
#> $modify_fn
#> function (original_prompt_text, llm_provider = NULL, http_list = NULL) 
#> {
#>     if (position == "after") {
#>         paste(original_prompt_text, text, sep = sep)
#>     }
#>     else {
#>         paste(text, original_prompt_text, sep = sep)
#>     }
#> }
#> <environment: 0x55f72fe3b650>
#> 
#> $extraction_fn
#> NULL
#> 
#> $validation_fn
#> NULL
#> 
#> $handler_fn
#> NULL
#> 
#> $parameter_fn
#> NULL
#> 
#> $name
#> [1] "add_text"
#> 
#> attr(,"class")
#> [1] "prompt_wrap"
#> 
#> [[2]]
#> $type
#> [1] "unspecified"
#> 
#> $modify_fn
#> function (original_prompt_text, llm_provider = NULL, http_list = NULL) 
#> {
#>     if (!add_instruction_to_prompt) {
#>         return(original_prompt_text)
#>     }
#>     glue::glue("{original_prompt_text}\n\n{instruction}")
#> }
#> <environment: 0x55f72fe40990>
#> 
#> $extraction_fn
#> function (x, llm_provider = NULL, http_list = NULL) 
#> {
#>     extracted <- suppressWarnings(as.numeric(x))
#>     if (is.na(extracted)) {
#>         return(llm_feedback(instruction))
#>     }
#>     return(extracted)
#> }
#> <environment: 0x55f72fe42a00>
#> 
#> $validation_fn
#> function (x, llm_provider = NULL, http_list = NULL) 
#> {
#>     if (x != floor(x)) {
#>         return(llm_feedback(instruction))
#>     }
#>     if (!is.null(min) && x < min) {
#>         return(llm_feedback(glue::glue("The number should be greater than or equal to {min}.")))
#>     }
#>     if (!is.null(max) && x > max) {
#>         return(llm_feedback(glue::glue("The number should be less than or equal to {max}.")))
#>     }
#>     return(TRUE)
#> }
#> <environment: 0x55f72fe44ae0>
#> 
#> $handler_fn
#> NULL
#> 
#> $parameter_fn
#> NULL
#> 
#> $name
#> [1] "answer_as_integer"
#> 
#> attr(,"class")
#> [1] "prompt_wrap"
#> 

# Construct prompt text
prompt$construct_prompt_text()
#> Hi
#> 
#> What is 5 + 5?
#> 
#> You must answer with only an integer (use no other characters).

# Configure chat history (affecting also the base prompt)
chat_history <- data.frame(
  role = c("user", "assistant", "user"),
  content = c("What is 5 + 5?", "10", "And what is 5 + 6?")
)
prompt$set_chat_history(chat_history)

# Get chat history
prompt$get_chat_history()
#> # A tibble: 3 × 3
#>   role      content                                                  tool_result
#>   <chr>     <chr>                                                    <lgl>      
#> 1 user      "What is 5 + 5?"                                         FALSE      
#> 2 assistant "10"                                                     FALSE      
#> 3 user      "And what is 5 + 6?\n\nWhat is 5 + 5?\n\nYou must answe… NA