Extract only prompt wraps and reorder them in order of operations

This function extracts only prompt wraps from a tidyprompt object and reorders them. The order of operations is as follows:

"Unspecified"

"Mode"

"Tool"

Usage

get_prompt_wraps_ordered(tidyprompt)

Arguments

tidyprompt

A tidyprompt object

Value

A list of prompt wraps from the tidyprompt, reordered in order of operations

See also

Examples

tidyprompt("Hi!")
#> <tidyprompt>
#> base prompt:
#> > Hi! 
#> Use '<tidyprompt>$base_prompt' to show the base prompt text.
#> Use '<tidyprompt> |> construct_prompt_text()' to get the full prompt text.
#> 

# Add to a tidyprompt using a prompt_wrap():
tidyprompt("Hi!") |>
  add_text("How are you?")
#> <tidyprompt>
#> The base prompt is modified by a wrapper function, resulting in:
#> > Hi!
#> > 
#> > How are you? 
#> Use '<tidyprompt>$prompt_wraps' to show the wrapper functions.
#> Use '<tidyprompt>$base_prompt' to show the base prompt text.
#> Use '<tidyprompt> |> construct_prompt_text()' to get the full prompt text.
#> 

# Strings can be input for prompt wraps; therefore,
#   a call to tidyprompt() is not necessary:
"Hi" |>
  add_text("How are you?")
#> <tidyprompt>
#> The base prompt is modified by a wrapper function, resulting in:
#> > Hi
#> > 
#> > How are you? 
#> Use '<tidyprompt>$prompt_wraps' to show the wrapper functions.
#> Use '<tidyprompt>$base_prompt' to show the base prompt text.
#> Use '<tidyprompt> |> construct_prompt_text()' to get the full prompt text.
#> 

# Example of adding extraction & validation with a prompt_wrap():
prompt <- "Hi" |>
  add_text("What is 5 + 5?") |>
  answer_as_integer()

if (FALSE) { # \dontrun{
  # tidyprompt objects are evaluated by send_prompt(), which will
  #   handle construct the prompt text, send it to the LLM provider,
  #   and apply the extraction and validation functions from the tidyprompt object
  prompt |>
    send_prompt(llm_provider_ollama())
  # --- Sending request to LLM provider (llama3.1:8b): ---
  #   Hi
  #
  #   What is 5 + 5?
  #
  #   You must answer with only an integer (use no other characters).
  # --- Receiving response from LLM provider: ---
  #   10
  # [1] 10

  # See prompt_wrap() and send_prompt() for more details
} # }

# `tidyprompt` objects may be validated with these helpers:
valid_tidyprompt <- validate_tidyprompt(prompt) # Returns self if valid
is_valid <- is_tidyprompt(prompt) # Returns TRUE if valid

# Get base prompt text
get_base_prompt(prompt)
#> [1] "Hi"

# Get all prompt wraps
get_prompt_wraps(prompt)
#> [[1]]
#> $type
#> [1] "unspecified"
#> 
#> $modify_fn
#> function (original_prompt_text) 
#> {
#>     if (position == "after") {
#>         paste(original_prompt_text, text, sep = sep)
#>     }
#>     else {
#>         paste(text, original_prompt_text, sep = sep)
#>     }
#> }
#> <bytecode: 0x556d3b8d3670>
#> <environment: 0x556d39cde190>
#> 
#> $extraction_fn
#> NULL
#> 
#> $validation_fn
#> NULL
#> 
#> attr(,"class")
#> [1] "prompt_wrap"
#> 
#> [[2]]
#> $type
#> [1] "unspecified"
#> 
#> $modify_fn
#> function (original_prompt_text) 
#> {
#>     if (!add_instruction_to_prompt) {
#>         return(original_prompt_text)
#>     }
#>     glue::glue("{original_prompt_text}\n\n{instruction}")
#> }
#> <bytecode: 0x556d38cc0640>
#> <environment: 0x556d39cdd898>
#> 
#> $extraction_fn
#> function (x) 
#> {
#>     extracted <- suppressWarnings(as.numeric(x))
#>     if (is.na(extracted)) {
#>         return(llm_feedback(instruction))
#>     }
#>     return(extracted)
#> }
#> <bytecode: 0x556d38cbbc68>
#> <environment: 0x556d39cdd898>
#> 
#> $validation_fn
#> function (x) 
#> {
#>     if (x != floor(x)) {
#>         return(llm_feedback(instruction))
#>     }
#>     if (!is.null(min) && x < min) {
#>         return(llm_feedback(glue::glue("The number should be greater than or equal to {min}.")))
#>     }
#>     if (!is.null(max) && x > max) {
#>         return(llm_feedback(glue::glue("The number should be less than or equal to {max}.")))
#>     }
#>     return(TRUE)
#> }
#> <bytecode: 0x556d38cbc4b8>
#> <environment: 0x556d39cdd898>
#> 
#> attr(,"class")
#> [1] "prompt_wrap"
#> 

# Get ordered prompt wraps (by type: tool, mode, unspecified)
get_prompt_wraps_ordered(prompt)
#> [[1]]
#> $type
#> [1] "unspecified"
#> 
#> $modify_fn
#> function (original_prompt_text) 
#> {
#>     if (position == "after") {
#>         paste(original_prompt_text, text, sep = sep)
#>     }
#>     else {
#>         paste(text, original_prompt_text, sep = sep)
#>     }
#> }
#> <bytecode: 0x556d3b8d3670>
#> <environment: 0x556d39cde190>
#> 
#> $extraction_fn
#> NULL
#> 
#> $validation_fn
#> NULL
#> 
#> attr(,"class")
#> [1] "prompt_wrap"
#> 
#> [[2]]
#> $type
#> [1] "unspecified"
#> 
#> $modify_fn
#> function (original_prompt_text) 
#> {
#>     if (!add_instruction_to_prompt) {
#>         return(original_prompt_text)
#>     }
#>     glue::glue("{original_prompt_text}\n\n{instruction}")
#> }
#> <bytecode: 0x556d38cc0640>
#> <environment: 0x556d39cdd898>
#> 
#> $extraction_fn
#> function (x) 
#> {
#>     extracted <- suppressWarnings(as.numeric(x))
#>     if (is.na(extracted)) {
#>         return(llm_feedback(instruction))
#>     }
#>     return(extracted)
#> }
#> <bytecode: 0x556d38cbbc68>
#> <environment: 0x556d39cdd898>
#> 
#> $validation_fn
#> function (x) 
#> {
#>     if (x != floor(x)) {
#>         return(llm_feedback(instruction))
#>     }
#>     if (!is.null(min) && x < min) {
#>         return(llm_feedback(glue::glue("The number should be greater than or equal to {min}.")))
#>     }
#>     if (!is.null(max) && x > max) {
#>         return(llm_feedback(glue::glue("The number should be less than or equal to {max}.")))
#>     }
#>     return(TRUE)
#> }
#> <bytecode: 0x556d38cbc4b8>
#> <environment: 0x556d39cdd898>
#> 
#> attr(,"class")
#> [1] "prompt_wrap"
#> 

# Construct prompt text
construct_prompt_text(prompt)
#> Hi
#> 
#> What is 5 + 5?
#> 
#> You must answer with only an integer (use no other characters).

# Get extraction and validation functions
get_extractions_and_validations(prompt)
#> $extractions
#> $extractions[[1]]
#> function (x) 
#> {
#>     extracted <- suppressWarnings(as.numeric(x))
#>     if (is.na(extracted)) {
#>         return(llm_feedback(instruction))
#>     }
#>     return(extracted)
#> }
#> <bytecode: 0x556d38cbbc68>
#> <environment: 0x556d39cdd898>
#> 
#> 
#> $validations
#> $validations[[1]]
#> function (x) 
#> {
#>     if (x != floor(x)) {
#>         return(llm_feedback(instruction))
#>     }
#>     if (!is.null(min) && x < min) {
#>         return(llm_feedback(glue::glue("The number should be greater than or equal to {min}.")))
#>     }
#>     if (!is.null(max) && x > max) {
#>         return(llm_feedback(glue::glue("The number should be less than or equal to {max}.")))
#>     }
#>     return(TRUE)
#> }
#> <bytecode: 0x556d38cbc4b8>
#> <environment: 0x556d39cdd898>
#> 
#>