Create a new Ollama llm_provider() instance

Create a new Ollama llm_provider() instance

Usage

llm_provider_ollama(
  parameters = list(model = "llama3.1:8b", stream = getOption("tidyprompt.stream", TRUE)),
  verbose = getOption("tidyprompt.verbose", TRUE),
  url = "http://localhost:11434/api/chat"
)

Arguments

parameters

A named list of parameters. Currently the following parameters are required:

model: The name of the model to use

stream: A logical indicating whether the API should stream responses Additional parameters may be passed by adding them to the parameters list; these parameters will be passed to the Ollama API via the body of the POST request. Options specifically can be set with the $set_options function (e.g., ollama$set_options(list(temperature = 0.8))). See available options at https://ollama.com/docs/api/chat

verbose

A logical indicating whether the interaction with the LLM provider should be printed to the console

url

The URL to the Ollama API

Value

A new llm_provider() object for use of the Ollama API

See also