Create a new OpenRouter llm_provider instance

Create a new OpenRouter llm_provider instance

Usage

llm_provider_openrouter(
  parameters = list(model = "qwen/qwen-2.5-7b-instruct", stream =
    getOption("tidyprompt.stream", TRUE)),
  verbose = getOption("tidyprompt.verbose", TRUE),
  url = "https://openrouter.ai/api/v1/chat/completions",
  api_key = Sys.getenv("OPENROUTER_API_KEY")
)

Arguments

parameters

A named list of parameters. Currently the following parameters are required:

model: The name of the model to use

stream: A logical indicating whether the API should stream responses Additional parameters are appended to the request body; see the OpenRouter API documentation for more information: https://openrouter.ai/docs/parameters

verbose

A logical indicating whether the interaction with the LLM provider should be printed to the console.

url

The URL to the OpenRouter API

api_key

The API key to use for authentication with the OpenRouter API

Value

A new llm_provider object for use of the OpenRouter API

See also