Tidyprompt R6 Class

A tidyprompt object contains a base prompt and a list of prompt_wrap() objects. It provides structured methods to modify the prompt while simultaneously adding logic to extract from and validate the LLM response. Besides a base prompt, a tidyprompt object may contain a system prompt and a chat history which precede the base prompt.

See also

Public fields

base_prompt

The base prompt string. The base prompt be modified by prompt wraps during construct_prompt_text(); the modified prompt text will be used as the final message of role 'user' during send_prompt()

system_prompt

A system prompt string. This will be added at the start of the chat history as role 'system' during send_prompt()

Methods

Public methods

tidyprompt-class$new()

tidyprompt-class$is_valid()

tidyprompt-class$add_prompt_wrap()

tidyprompt-class$get_prompt_wraps()

tidyprompt-class$construct_prompt_text()

tidyprompt-class$set_chat_history()

tidyprompt-class$get_chat_history()

tidyprompt-class$clone()

Method new()

Initialize a tidyprompt object

Usage

tidyprompt-class$new(input)

Arguments

input

A string, a chat history, a list containing a chat history under key '$chat_history', or a tidyprompt object

Details

Different types of input are accepted for initialization of a tidyprompt object:

A single character string. This will be used as the base prompt

A dataframe which is a valid chat history (see chat_history())

A list containing a valid chat history under '$chat_history' (e.g., a result from send_prompt() when using 'return_mode' = "full")

A tidyprompt object. This will be checked for validity and, if valid, the fields are copied to the object which is returned from this method

When passing a dataframe or list with a chat history, the last row of the chat history must have role 'user'; this row will be used as the base prompt. If the first row of the chat history has role 'system', it will be used as the system prompt.

Returns

A tidyprompt object

Method is_valid()

Check if the tidyprompt object is valid.

Usage

tidyprompt-class$is_valid()

Returns

TRUE if valid, otherwise FALSE

Method add_prompt_wrap()

Add a prompt_wrap() to the tidyprompt object.

Usage

tidyprompt-class$add_prompt_wrap(prompt_wrap)

Arguments

prompt_wrap

A prompt_wrap() object

Returns

The updated tidyprompt object

Method get_prompt_wraps()

Get list of prompt_wrap() objects from the tidyprompt object.

Usage

tidyprompt-class$get_prompt_wraps(
  order = c("default", "modification", "evaluation")
)

Arguments

order

The order to return the wraps. Options are:

"default": as originally added to the object

"modification": as ordered for modification of the base prompt; ordered by type: check, unspecified, mode, tool, break. This is the order in which prompt wraps are applied during construct_prompt_text()

"evaluation": ordered for evaluation of the LLM response; ordered by type: tool, mode, break, unspecified, check. This is the order in which wraps are applied to the LLM output during send_prompt()

Returns

A list of prompt_wrap() objects.

Method construct_prompt_text()

Construct the complete prompt text.

Usage

tidyprompt-class$construct_prompt_text(llm_provider = NULL)

Arguments

llm_provider

Optional llm_provider object. This may sometimes affect the prompt text construction

Returns

A string representing the constructed prompt text

Method set_chat_history()

This function sets the chat history for the tidyprompt object. The chat history will also set the base prompt and system prompt (the last message of the chat history should be of role 'user' and will be used as the base prompt; the first message of the chat history may be of the role 'system' and will then be used as the system prompt). This may be useful when one wants to change the base prompt, system prompt, and chat history of a tidyprompt object while retaining other fields like the prompt wraps.

Usage

tidyprompt-class$set_chat_history(chat_history)

Arguments

chat_history

A valid chat history (see chat_history())

Returns

The updated tidyprompt object

Method get_chat_history()

This function gets the chat history of the tidyprompt object. The chat history is constructed from the base prompt, system prompt, and chat history field. The returned object will be the chat history with the system prompt as the first message with role 'system' and the the base prompt as the last message with role 'user'.

Usage

tidyprompt-class$get_chat_history()

Returns

A dataframe containing the chat history

Method clone()

The objects of this class are cloneable with this method.

Usage

tidyprompt-class$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

tidyprompt("Hi!")
#> <tidyprompt>
#> The base prompt is not modified by prompt wraps:
#> > Hi! 
#> Use 'x$base_prompt' to show the base prompt text.
#> Use 'x$construct_prompt_text()' to get the full prompt text.
#> 

# Add to a tidyprompt using a prompt wrap:
tidyprompt("Hi!") |>
  add_text("How are you?")
#> <tidyprompt>
#> The base prompt is modified by a prompt wrap, resulting in:
#> > Hi!
#> > 
#> > How are you? 
#> Use 'x$base_prompt' to show the base prompt text.
#> Use 'x$construct_prompt_text()' to get the full prompt text.
#> Use 'get_prompt_wraps(x)' to show the prompt wraps.
#> 

# Strings can be input for prompt wraps; therefore,
#   a call to tidyprompt() is not necessary:
"Hi" |>
  add_text("How are you?")
#> <tidyprompt>
#> The base prompt is modified by a prompt wrap, resulting in:
#> > Hi
#> > 
#> > How are you? 
#> Use 'x$base_prompt' to show the base prompt text.
#> Use 'x$construct_prompt_text()' to get the full prompt text.
#> Use 'get_prompt_wraps(x)' to show the prompt wraps.
#> 

# Example of adding extraction & validation with a prompt_wrap():
prompt <- "Hi" |>
  add_text("What is 5 + 5?") |>
  answer_as_integer()

if (FALSE) { # \dontrun{
  # tidyprompt objects are evaluated by send_prompt(), which will
  #   handle construct the prompt text, send it to the LLM provider,
  #   and apply the extraction and validation functions from the tidyprompt object
  prompt |>
    send_prompt(llm_provider_ollama())
  # --- Sending request to LLM provider (llama3.1:8b): ---
  #   Hi
  #
  #   What is 5 + 5?
  #
  #   You must answer with only an integer (use no other characters).
  # --- Receiving response from LLM provider: ---
  #   10
  # [1] 10

  # See prompt_wrap() and send_prompt() for more details
} # }

# `tidyprompt` objects may be validated with these helpers:
is_tidyprompt(prompt) # Returns TRUE if input is a valid tidyprompt object
#> [1] TRUE

# Get base prompt text
prompt$base_prompt
#> [1] "Hi"

# Get all prompt wraps
prompt$get_prompt_wraps()
#> [[1]]
#> $type
#> [1] "unspecified"
#> 
#> $modify_fn
#> function (original_prompt_text, llm_provider = NULL, http_list = NULL) 
#> {
#>     if (position == "after") {
#>         paste(original_prompt_text, text, sep = sep)
#>     }
#>     else {
#>         paste(text, original_prompt_text, sep = sep)
#>     }
#> }
#> <environment: 0x563f242f22e0>
#> 
#> $extraction_fn
#> NULL
#> 
#> $validation_fn
#> NULL
#> 
#> $handler_fn
#> NULL
#> 
#> $parameter_fn
#> NULL
#> 
#> $name
#> [1] "add_text"
#> 
#> attr(,"class")
#> [1] "prompt_wrap"
#> 
#> [[2]]
#> $type
#> [1] "unspecified"
#> 
#> $modify_fn
#> function (original_prompt_text, llm_provider = NULL, http_list = NULL) 
#> {
#>     if (!add_instruction_to_prompt) {
#>         return(original_prompt_text)
#>     }
#>     glue::glue("{original_prompt_text}\n\n{instruction}")
#> }
#> <environment: 0x563f242f36d8>
#> 
#> $extraction_fn
#> function (x, llm_provider = NULL, http_list = NULL) 
#> {
#>     extracted <- suppressWarnings(as.numeric(x))
#>     if (is.na(extracted)) {
#>         return(llm_feedback(instruction))
#>     }
#>     return(extracted)
#> }
#> <environment: 0x563f242f5748>
#> 
#> $validation_fn
#> function (x, llm_provider = NULL, http_list = NULL) 
#> {
#>     if (x != floor(x)) {
#>         return(llm_feedback(instruction))
#>     }
#>     if (!is.null(min) && x < min) {
#>         return(llm_feedback(glue::glue("The number should be greater than or equal to {min}.")))
#>     }
#>     if (!is.null(max) && x > max) {
#>         return(llm_feedback(glue::glue("The number should be less than or equal to {max}.")))
#>     }
#>     return(TRUE)
#> }
#> <environment: 0x563f24254138>
#> 
#> $handler_fn
#> NULL
#> 
#> $parameter_fn
#> NULL
#> 
#> $name
#> [1] "answer_as_integer"
#> 
#> attr(,"class")
#> [1] "prompt_wrap"
#> 

# Construct prompt text
prompt$construct_prompt_text()
#> Hi
#> 
#> What is 5 + 5?
#> 
#> You must answer with only an integer (use no other characters).

# Configure chat history (affecting also the base prompt)
chat_history <- data.frame(
  role = c("user", "assistant", "user"),
  content = c("What is 5 + 5?", "10", "And what is 5 + 6?")
)
prompt$set_chat_history(chat_history)

# Get chat history
prompt$get_chat_history()
#> # A tibble: 3 × 3
#>   role      content                                                  tool_result
#>   <chr>     <chr>                                                    <lgl>      
#> 1 user      "What is 5 + 5?"                                         FALSE      
#> 2 assistant "10"                                                     FALSE      
#> 3 user      "And what is 5 + 6?\n\nWhat is 5 + 5?\n\nYou must answe… NA