tidyprompt: Prompt and empower your LLM, the tidy way

The 'tidyprompt' package allows users to prompt and empower their large language models (LLMs) in a tidy way. It provides a framework to construct LLM prompts using 'tidyverse'-inspired piping syntax, with a library of pre-built prompt wrappers and the option to build custom ones. Additionally, it supports structured LLM output extraction and validation, with automatic feedback and retries if necessary. Moreover, it enables specific LLM reasoning modes, autonomous R function calling for LLMs, and compatibility with any LLM provider.

See also

Author

Maintainer: Luka Koning l.koning@kennispunttwente.nl [copyright holder]

Authors:

Tjark Van de Merwe t.vandemerwe@kennispunttwente.nl [copyright holder]