Wrap a prompt with empowering functions

This function takes a single string or a tidyprompt-class object and adds a new prompt wrap to it. A prompt wrap is a set of functions that modify the prompt text, extract a value from the LLM response, and validate the extracted value. The functions are used to ensure that the prompt and LLM response is in the correct format and meets the specified criteria; they may also be used to provide the LLM with feedback or additional information, like the result of a tool call.

Usage

prompt_wrap(
  prompt,
  modify_fn = NULL,
  extraction_fn = NULL,
  validation_fn = NULL,
  handler_fn = NULL,
  parameter_fn = NULL,
  type = c("unspecified", "mode", "tool", "break"),
  name = NULL
)

Arguments

prompt

A single string or a tidyprompt-class object

modify_fn

A function that takes the previous prompt text (as first argument) and returns the new prompt text

extraction_fn

A function that takes the LLM response (as first argument) and attempts to extract a value from it. Upon succesful extraction, the function should return the extracted value. If the extraction fails, the function should return a llm_feedback() message to initiate a retry. A llm_break() can be returned to break the extraction and validation loop, ending send_prompt()

validation_fn

A function that takes the (extracted) LLM response (as first argument) and attempts to validate it. Upon succesful validation, the function should return TRUE. If the validation fails, the function should return a llm_feedback() message to initiate a retry. A llm_break() can be returned to break the extraction and validation loop, ending send_prompt() object llm_break() can be returned to break the extraction and validation loop

handler_fn

A function that takes a 'completion' object (as returned by llm_provider$complete_chat()) and a (potentially modified) completion object. This can be used for advanced side effects, like logging, or native tool calling, or keeping track of token usage. See llm_provider-class for more information; handler_fn is attached to the llm_provider-class object that is being used. For example usage, see source code of answer_using_tools()

parameter_fn

A function that takes the llm_provider-class object which is being used with send_prompt() and returns a named list of parameters to be set in the llm_provider-class object via llm_provider$set_parameters(). This can be used to configure specific parameters of the llm_provider-class object when evaluating the prompt. For example, answer_as_json() may set different parameters for different APIs related to JSON output. This function is typically only used with advanced prompt wraps that require specific settings in the llm_provider-class object

type

The type of prompt wrap; one of 'unspecified', 'mode', 'tool', or 'break'. Types are used to determine the order in which prompt wraps are applied. When constructing the prompt text, prompt wraps are applied to the base prompt in the following order: 'unspecified', 'break', 'mode', 'tool'. When evaluating the LLM response and applying extraction and validation functions, prompt wraps are applied in the reverse order: 'tool', 'mode', 'break', 'unspecified'. Order among the same type is preserved in the order they were added to the prompt. Example of a tool is answer_using_tools(); example of a mode is answer_by_react(). Example of a break is quit_if(). Most other prompt wraps will be 'unspecified', like answer_as_regex() or add_text()

name

An optional name for the prompt wrap. This can be used to identify the prompt wrap in the tidyprompt-class object

Value

A tidyprompt-class object with the prompt_wrap() appended to it

Details

For advanced use, modify_fn, extraction_fn, and validation_fn may take 'llm_provider' (as used with send_prompt()) as second argument, and the http_list (a list of all HTTP requests made during send_prompt()) as third argument. This is not required, but can be useful for more complex prompt wraps which require additional information about the LLM provider or the HTTP requests made.

'modify_fn', 'extraction_fn', 'validation_fn', 'handler_fn', and 'parameter_fn' may also access the tidyprompt-class object that they are a part of through 'self$...' which is attached to their environment. For instance, other prompt wraps can be accessed through 'self$get_prompt_wraps()'

See also

Examples

# A custom prompt_wrap may be created during piping
prompt <- "Hi there!" |>
  prompt_wrap(
    modify_fn = function(base_prompt) {
      paste(base_prompt, "How are you?", sep = "\n\n")
    }
  )
prompt
#> <tidyprompt>
#> The base prompt is modified by a prompt wrap, resulting in:
#> > Hi there!
#> > 
#> > How are you? 
#> Use 'x$base_prompt' to show the base prompt text.
#> Use 'x$construct_prompt_text()' to get the full prompt text.
#> Use 'get_prompt_wraps(x)' to show the prompt wraps.
#> 

# (Shorter notation of the above:)
prompt <- "Hi there!" |>
  prompt_wrap(\(x) paste(x, "How are you?", sep = "\n\n"))

# It may often be preferred to make a function which takes a prompt and
#   returns a wrapped prompt:
my_prompt_wrap <- function(prompt) {
  modify_fn <- function(base_prompt) {
    paste(base_prompt, "How are you?", sep = "\n\n")
  }

  prompt_wrap(prompt, modify_fn)
}
prompt <- "Hi there!" |>
  my_prompt_wrap()

# For more advanced examples, take a look at the source code of the
#   pre-built prompt wraps in the tidyprompt package, like
#   answer_as_boolean, answer_as_integer, add_tools, answer_as_code, etc.
# Below is the source code for the 'answer_as_integer' prompt wrap function:

#' Make LLM answer as an integer (between min and max)
#'
#' @param prompt A single string or a [tidyprompt()] object
#' @param min (optional) Minimum value for the integer
#' @param max (optional) Maximum value for the integer
#' @param add_instruction_to_prompt (optional) Add instruction for replying
#' as an integer to the prompt text. Set to FALSE for debugging if extractions/validations
#' are working as expected (without instruction the answer should fail the
#' validation function, initiating a retry)
#'
#' @return A [tidyprompt()] with an added [prompt_wrap()] which
#' will ensure that the LLM response is an integer.
#'
#' @export
#'
#' @example inst/examples/answer_as_integer.R
#'
#' @family pre_built_prompt_wraps
#' @family answer_as_prompt_wraps
answer_as_integer <- function(
    prompt,
    min = NULL,
    max = NULL,
    add_instruction_to_prompt = TRUE
) {
  instruction <- "You must answer with only an integer (use no other characters)."

  if (!is.null(min) && !is.null(max)) {
    instruction <- paste(instruction, glue::glue(
      "Enter an integer between {min} and {max}."
    ))
  } else if (!is.null(min)) {
    instruction <- paste(instruction, glue::glue(
      "Enter an integer greater than or equal to {min}."
    ))
  } else if (!is.null(max)) {
    instruction <- paste(instruction, glue::glue(
      "Enter an integer less than or equal to {max}."
    ))
  }

  modify_fn <- function(original_prompt_text) {
    if (!add_instruction_to_prompt) {
      return(original_prompt_text)
    }

    glue::glue("{original_prompt_text}\n\n{instruction}")
  }

  extraction_fn <- function(x) {
    extracted <- suppressWarnings(as.numeric(x))
    if (is.na(extracted)) {
      return(llm_feedback(instruction))
    }
    return(extracted)
  }

  validation_fn <- function(x) {
    if (x != floor(x)) { # Not a whole number
      return(llm_feedback(instruction))
    }

    if (!is.null(min) && x < min) {
      return(llm_feedback(glue::glue(
        "The number should be greater than or equal to {min}."
      )))
    }
    if (!is.null(max) && x > max) {
      return(llm_feedback(glue::glue(
        "The number should be less than or equal to {max}."
      )))
    }
    return(TRUE)
  }

  prompt_wrap(
    prompt,
    modify_fn, extraction_fn, validation_fn,
    name = "answer_as_integer"
  )
}