Tidyprompt R6 Class

A Tidyprompt is an object containing a base prompt string and a list of prompt_wrap objects. It provides structured methods to modify the prompt, validate responses, and chain modifications, ensuring robust interaction with LLMs.

Details

This class includes methods for adding prompt wraps, constructing prompt text, and validating the object. It allows chaining modifications to the prompt and validating responses, ensuring a seamless flow in prompt design.

See also

Public fields

base_prompt

The base prompt string

system_prompt

A system prompt string

Methods

Public methods

tidyprompt-class$new()

tidyprompt-class$is_valid()

tidyprompt-class$add_prompt_wrap()

tidyprompt-class$get_prompt_wraps()

tidyprompt-class$construct_prompt_text()

tidyprompt-class$clone()

Method new()

Initialize a Tidyprompt object.

Usage

tidyprompt-class$new(input)

Arguments

input

A character string representing the base prompt.

Returns

A Tidyprompt object.

Method is_valid()

Check if the Tidyprompt object is valid.

Usage

tidyprompt-class$is_valid()

Returns

TRUE if valid, otherwise FALSE.

Method add_prompt_wrap()

Add a prompt_wrap object to the Tidyprompt.

Usage

tidyprompt-class$add_prompt_wrap(prompt_wrap)

Arguments

prompt_wrap

A prompt_wrap object.

Returns

The updated Tidyprompt object (invisible).

Method get_prompt_wraps()

Get the list of prompt wraps.

Usage

tidyprompt-class$get_prompt_wraps(
  order = c("default", "modification", "evaluation")
)

Arguments

order

The order to return the wraps. Options are:

"default": as added.

"modification": ordered by modification stages.

"evaluation": ordered by evaluation stages.

Returns

A list of prompt_wrap objects.

Method construct_prompt_text()

Construct the complete prompt text.

Usage

tidyprompt-class$construct_prompt_text(llm_provider = NULL)

Arguments

llm_provider

Optional LLM provider for context.

Returns

A character string representing the constructed prompt text.

Method clone()

The objects of this class are cloneable with this method.

Usage

tidyprompt-class$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.